# backend/services/ai-chatbot/openapi.yaml
swagger: '2.0'
info:
  title: Grassroots Football Management API
  description: API Gateway for Grassroots Football Management microservices
  version: 1.0.0
host: gfm-1ykrsotxghm8m.apigateway.grassroots-football-management.cloud.goog
...
schemes:
  - https
securityDefinitions:
  firebase:
    type: oauth2
    flow: "implicit"
    authorizationUrl: ""
    x-google-issuer: "https://securetoken.google.com/grassroots-football-management"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "grassroots-football-management"
security:
  - firebase: []
paths:
...
  /fixture/lineups:
    post:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Save initial lineups for a match
      operationId: saveLineups
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division, homeTeamLineup, awayTeamLineup]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              homeTeamLineup:
                type: object
              awayTeamLineup:
                type: object
      responses:
        '201':
          description: Lineups saved successfully.
    get:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve lineups for a match
      operationId: getLineups
      parameters:
        - in: query
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Lineups retrieved successfully.
    patch:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update existing lineups for a match
      operationId: updateLineups
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              homeTeamLineup:
                type: object
              awayTeamLineup:
                type: object
      responses:
        '200':
          description: Lineups updated successfully.
    delete:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete lineups for a match
      operationId: deleteLineups
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
      responses:
        '200':
          description: Lineups deleted successfully.
  /fixture/events:
    post:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Add a new event to a match
      operationId: addEvent
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division, event]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              event:
                type: object
      responses:
        '201':
          description: Event added successfully.
    get:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve all events for a match
      operationId: getEvents
      parameters:
        - in: query
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Events retrieved successfully.
    patch:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update an existing match event
      operationId: updateEvent
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division, eventId, event]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              eventId:
                type: string
              event:
                type: object
      responses:
        '200':
          description: Event updated successfully.
    delete:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete a specific event from a match
      operationId: deleteEvent
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division, eventId]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              eventId:
                type: string
      responses:
        '200':
          description: Event deleted successfully.
  /fixture/results:
    post:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Save initial match result
      operationId: saveResult
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division, homeScore, awayScore]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              homeScore:
                type: integer
              awayScore:
                type: integer
      responses:
        '201':
          description: Result saved successfully.
    get:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve match result
      operationId: getResult
      parameters:
        - in: query
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Result retrieved successfully.
    patch:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update match result
      operationId: updateResult
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division]
            properties:
              homeScore:
                type: integer
              awayScore:
                type: integer
      responses:
        '200':
          description: Result updated successfully.
    delete:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete match result
      operationId: deleteResult
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division]
            properties:
              matchId:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
      responses:
        '200':
          description: Result deleted successfully.
        '404':
          description: Result not found.
        '500':
          description: Internal server error.
  /fixture/player:
    post:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Submit or overwrite player rating for a match
      description: Submits or overwrites player rating with dynamically accepted fields.
      operationId: submitPlayerRating
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [clubName, ageGroup, division, matchId, playerEmail]
            properties:
              clubName:
                type: string
                description: The name of the club.
              ageGroup:
                type: string
                description: The age group (e.g., U12).
              division:
                type: string
                description: The division (e.g., Premier).
              matchId:
                type: string
                description: The match ID.
              playerEmail:
                type: string
                description: The email of the player being rated.
            additionalProperties:  # This allows any other rating fields to be dynamically accepted.
              type: number
              description: Any additional player rating metric (e.g., pace, passingAccuracy, teamwork).
      responses:
        '201':
          description: Player rating submitted successfully.
        '400':
          description: Missing required fields or invalid input.
        '500':
          description: Internal server error.

    get:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Retrieve a player rating for a match
      description: Fetches a player rating for a specific match.
      operationId: getPlayerRating
      produces:
        - application/json
      parameters:
        - in: query
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: List of player ratings for the match.
          schema:
            type: array
            items:
              type: object
              properties:
                playerEmail:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                additionalProperties:
                  type: number
        '500':
          description: Internal server error.

    patch:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update player rating fields for a match
      description: Allows partial updates to player ratings, dynamically accepting new fields.
      operationId: updatePlayerRating
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [clubName, ageGroup, division, matchId, playerEmail]
            properties:
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              matchId:
                type: string
              playerEmail:
                type: string
            additionalProperties:
              type: number
              description: Any additional player rating metric to update.
      responses:
        '200':
          description: Player rating updated successfully.
        '400':
          description: No valid fields provided for update.
        '500':
          description: Internal server error.

    delete:
      x-google-backend:
        address: https://microservice-match-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-match-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete player rating for a match
      description: Deletes a player’s rating for a specific match.
      operationId: deletePlayerRating
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [clubName, ageGroup, division, matchId, playerEmail]
            properties:
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              matchId:
                type: string
              playerEmail:
                type: string
      responses:
        '200':
          description: Player rating deleted successfully.
        '500':
          description: Internal server error.


  # Carpool Microservice Routes
...